pyproject.toml
[project]
name = "infiniteagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "uvicorn>=0.41.0",
]
